Add a function to repair triggers that were obfuscated by the photodiode signal
Over the course of the experiment, the behavior of a photodiode was changed. The signal of the diode overlays
all other trigger signals it coincides with, resulting in obfuscated trigger events, hidden/cancelled events,
and duplication in event series.
This function removes the photodiode signal from the event channels by replacing every photodiode signal
with the trigger signal it overlaid. If this results in identical consecutive signals/events, the first one
(which likely corresponds to the converted signal of the photodiode and is more accurate) is kept, while the
latter is discarded in order to have only one occurance of any given event per trial.