Overthink the approach to ICA
Having read up on the topic of high-pass filtering distoring decoding performance, were only filtering
a temporary data copy with a high high-passfilter (1Hz) to fit the ICA on. The ICA parameters are applied
to unfiltered raw data.
The estimation of rejection criteria via autoreject on artifical epochs is moved into the ICA function