Retrieve trial data not only for the left visual stimulus, but also for the second
This is done to demonstrate what hopefully is the hemifield difference between the two
stimulations. The general data structure does not support a representation of the
data in one structure easily though. Hence, this change introduces plenty of
duplication, in that all data structures now exist for 'left stimuli' and for
'right stimuli'