Add a hack: ensure linked flatteners to pass sampleshape attributes
Cave, this isn't an elegant way to solve this, but the only way we came up with.
When cross_val_multiscore clones (i.e., deepcopies) elements of the pipeline,
it reinstantiates the Flattener used and trained during dimensionality reduction.
The sampleshape thats stored during dimensionality reduction and needed during
later steps of the pipeline to reshape the data X back into the desired state
is lost during this process. Keeping the sampleshape parameter despite cloning
turned out to be tricky - its not an estimator where get_params() might have
helped. A process that worked is a hack: We implement our own __deepcopy__() to
have control over what cross_val_multiscore's 'clone()' does, and have our
__deepcopy__() not actually do a deepcopy. Instead of storing the sampleshape
as a class attribute, we place it in a dictionary. The dictionary is an attribute,
but it will point to the same dict always, ensuring that a clone of the Flattener
has access to the sampleshape trained in a different part of the pipeline.