Add custom pseudo-scorer to return activation patterns following Haufe et al, 2014
The scorer does not implement any scoring, but soley exists to retrieve
an internal attribute from the SRM estimator. This attribute can hold
activation patterns created from the subject specific transformation
matrices from the SRM model. This is slightly hacky, but the only easy
way to surface data from that deep in the pipeline